pyproject.toml
[project]
name = "rebalancer"
version = "0.1.0"
description = "Portfolio rebalancer — aggregate holdings, classify by region & category, rebalance to targets"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0",
    "flask-sqlalchemy>=3.1",
    "flask-migrate>=4.0",
    "psycopg2-binary>=2.9",
    "gunicorn>=22.0",
    "python-dotenv>=1.0",
    "openai>=1.0",
]
